Move regenerateAll to JenkinsJobUpdater
